Fixed warnings and updating rating to decimal.

+++ db/migrate/20190328211806_add_rating_to_agencies.rb
@@ -1,5 +1,5 @@
 class AddRatingToAgencies < ActiveRecord::Migration[5.2]
   def change
-    add_column :agencies, :rating, :integer
+    add_column :agencies, :rating, :decimal
   end
 end
